data-simulator/src/temperature_controller/requirements.txt
paho-mqtt
requests
ssl
